feat: extract a common route for notebook displaying
