+++ src/routes/(notebook)/+layout.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
-  import readme from "../../../README.md?raw";
-  import Notebook from "$lib/components/notebook/Notebook.svelte";
   import { Button } from "bits-ui";
 </script>
 
-<div class="m-4 max-w-[calc(100%-2rem)] w-auto self-center lg:m-7 md:m-6 sm:m-5 xl:m-8 2xl:max-w-4xl lg:max-w-2xl sm:max-w-xl xl:max-w-3xl [&>article]:(lg:text-3.75 xl:text-base)">
+<div class="m-4 w-[calc(100%-2rem)] self-center lg:(m-7 max-w-2xl) md:m-6 sm:(m-5 max-w-xl w-auto) xl:m-8 2xl:max-w-4xl xl:max-w-3xl [&>article]:(lg:text-3.75 xl:text-base)">
 
   <nav class="w-full flex flex-row items-center justify-between gap-2 text-sm lg:text-base [&>a:hover]:op-80 [&>a]:(op-50 transition)">
     <Button.Root href="/">Home</Button.Root>
@@ -13,5 +11,5 @@
 
   <hr class="invisible h-6">
 
-  <Notebook text={readme} />
+  <slot />
 </div>
