+++ src/routes/(notebook)/about/+page.server.ts
