+++ src/routes/(notebook)/about/+page.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+  import readme from "../../../../README.md?raw";
+  import Notebook from "$lib/components/notebook/Notebook.svelte";
+</script>
+
+<Notebook text={readme} />
