+++ src/routes/(notebook)/cpython/[...path]/+page.server.ts
